+++ packages/react/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `<Link>` uses `history.navigate`, not `history.update`.
+
 ## 1.0.0-beta.11
 
 * Update types to expect `Response` instead of `AnyResponse` from `@curi/core`.